requirements.txt
pdfminer.six
prompt_toolkit
requests
slugify
googlesearch
elasticsearch
bs4
nltk
